Lifted state up to App.jsx to handle route and view swapping. #2
